Changed default prototype for gethostname. Original: 57fbf6530620e1264d26550e8aee1e8d6ff3f477

+++ lal/packages/framedata/src/FrameStream.c
@@ -72,7 +72,7 @@
 #include <config.h>
 #include <unistd.h>
 #ifndef HAVE_GETHOSTNAME_PROTOTYPE
-int gethostname(char *name, size_t len);
+int gethostname(char *name, int len);
 #endif
 
 #include <FrameL.h>
